Provide backward compatibility for the instance ID changes in commit f1eeb4efb9d147a9f457e4602ab364d24d5a293c
In f1eeb4efb9d147a9f457e4602ab364d24d5a293c the VMID instance IDs of Pipe, Alphabet, and PagedInstanceList were replaced with UUID instance IDs to provide compatibility with Android, which has no VMID class. However, that commit did not properly address the backward-incompatible nature of that change.

Commit d04a3bc02c8c75c998e6783d720d2976a319af26 previously incremented the CURRENT_SERIAL_VERSION in Pipe, Alphabet, and PagedInstanceList. That commit can be used to re-serialize models serialized since f1eeb4efb9d147a9f457e4602ab364d24d5a293c with a properly-updated version number. That should be done prior to loading using this commit and subsequent.

This commit uses the version to conditionally either load a VMID and convert it to a UUID, or load a UUID directly.

The VMIDs are converted to UUIDs by MD5-hashing their string representation to a 128-bit value, which is then used to initialize the UUID. This is deterministic, meaning multiple instances of the same model will yield the same UUID, allowing duplicates to be detected and ignored (which appears to be the purpose of the instance ID fields in Pipe, Alphabet, and PagedInstanceList).

It should be noted that the `PagedInstanceList.readObject` was not reading in its own version field, apparently in error. The version is now read first and used for backward-compatible loading of the instance ID.

This approach should still maintain Android compatibility as it will only use VMID at runtime if it is attempting to deserialize an old model with the old serialization version. If this is avoided then Android should do alright, though this has not been tested as of yet.